pyproject.toml
[tool.poetry]
name = "fish"
version = "0.1.0"
description = ""
authors = ["liuxu <xu.liu@example.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
starlette = "^0.28.0"
uvicorn = "^0.22.0"
tortoise-orm = "^0.19.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
